URI Filtering Options
Added the ability to define a default page in the opts parameter that's
passed into the WebApplication object. If the default option is
supplied, any requests to '/' will be redirected to the value of
opts.default.
For more control on rerouting, the user can supply a function to
opts.uriFilter. The original uri is passed to the function and the
function's return value is used for the actual request.